+++ l1l2Penalty.jl
@@ -9,11 +9,11 @@ end
 
 function update_model(reg::L1L2Penalty, old_w::Float64, grad::Float64, eta::Float64) 
 	#eta: step-size
-	temp = eta*old_w - grad
+	temp::Float64 = eta*old_w - grad
 	if (abs(temp) < reg.lambda_l1)
-		return 0
+		return 0.0
 	else
-		ret = sign(temp) * (abs(temp) - reg.lambda_l1)
+		ret::Float64 = sign(temp) * (abs(temp) - reg.lambda_l1)
 		return ret / (eta + reg.lambda_l2)
 	end
 end
